Fix skipping the copy of empty dirs on tree copies
The CopyDir implementation did not copy empty directories which was
resulting in a dirty repo after the tree copy with non-cloned submodules
(empty dirs) was performed.

The issue resulted from calling `MkdirAll` with the parent directory of
the current directory under copy, followed by a `ReadDir` which would
result in an empty list of files/dirs, skipping the directory creation
(by skipping the recursive tree copy).

Signed-off-by: Fabio Utzig <utzig@apache.org>