downloader: Return base commit if no rc tag
downloader.LatestRc() infers the latest "rc" tag given a base string
(e.g., "mynewt_1_8_0_tag" might resolve to "mynewt_1_8_0_rc3_tag").

If there are no rc tags for the given base, this function used to return
"".  This is error prone and it requires the caller to stash the base
string until he has inspected the return value.

This commit changes this function so that it returns the base string
unmodified if there are no rc tags.