upgrade: Don't abort if repo references bad commit
Newt would abort the upgrade if a repo's dependency list specifies a
nonexistent commit.  This was bad because it means a repo maintainer
could break `newt upgrade` for all users by deleting a tag or by
introducing a typo to their `repository.yml` file.

Now newt just warns the user about the problem and proceeds with the
upgrade.