build.sh: Invoke sh as a login shell
This fixes #390

build.sh depends on the `go` executable.  This executable may not be in the
user's path when /bin/sh is invoked as a non-login shell (a non-login
shell does not execute the various "profile" scripts).  Since we don't
package `go` as part of the newt distribution, we should not assume it
is available to the build script in a non-login shell.