project: Lazily load package set
Don't load the package set until it is needed.

Loading the set of packages in a project is a slow operation.  It
requires opening all of the pkg.yml and syscfg.yml files in the project
tree.

Some newt commands (e.g., `info`, `upgrade`) don't require the package
set to be loaded.  This change speeds up those commands.