Add subpriorities to packages
This extends concept of subpriorities from config packages to all
"standard" packages, i.e. packages that are not target/config/app/bsp.

By default, package has subrpiority of 0 and this should not be changed
unless really necessary. An example where this can be useful are MCU
packages where we have a generic package for MCU family and another
package for specific MCU. In that case it may be useful to allow the
latter package override settings from the former package. Since both
packages are of the same (standard) type, without priorities it may not
be easy to do so.