toolchain: Link all archives with -whole-archive
Using -whole-archives guarantees that all objects from .a files are
included in build. It should not affect regular builds since we want to
include all objects from Mynewt packages anyway, but it helps to resolve
weak symbols properly if weak and non-weak counterparts are defined in
the same package/library/archive.