Add const syscfg values
This allows to define syscfg value that shall be const. A const syscfg
definition is simply a syscfg.vals setting placed in syscfg.defs
section.

An example of candicates for such vals are e.g. BLE_CONTROLLER, BLE_HOST
or BSP_* values that are defined to be used as conditionals in other
configuration files or in code. This change removes burden of adding a
full definition and possibly also restricions for syscfg we do not want
to be ever changed.