Fix overwriting newtrc settings with default value
newtrc settings are read properly, but then escape_shell value is
overwritten anyway by command line setting (or a default value) so it
was not really working as expected. Now newtrc settings are read at the
very beginning and then default value or value read from newtrc is used
as a default value for command line flag so it won't be overwritten.