Add support for expressions in syscfg values
This adds support for evaluating syscfg values as expressions. To make
it compatible with existing code, syscfg value is only evaluated as
expression if its type is explicitly set to "expr", i.e.:

syscfg.defs:
  FOO:
    description: ...
    type: expr
    value: 1 + 2 + 3

Following tokens are allowed in expressions:
- literals (integers and strings)
- identifiers (references to other syscfg values)
- parentheses
- binary operators (arthmetic, relational and boolean)
- unary operator (boolean negation)
- built-in function calls

Most of operators support only integer values. Strings are supported by
"==" and "!=" only.

Available built-in functions are:
- min(a,b) - returns lesser of "a" and "b"
- max(a,b) - returns greater of "a" and "b"
- in_range(v,a,b) - returns if "v" is inside [a,b] range
- clamp(v,a,b) - clamps "v" to be inside [a,b] range
- ite(v,a,b) - if-then-else, returns "a" if "v", otherwise returns "b"
- in_set(v,...) - returns if "v" is one of remaining arguments

Note: all arguments to built-in functions shall be integer only, except
for "a" and "b" in ite() and all arguments in in_set().