Add project cache
This adds cache for project packages. It can be enabled with
"project_cache: 1" in newtrc.yml.

Once enabled, packages list for each repo will be stored in .cache
directory and read from there on subsequent executions. This allows to
substantially strip time required to find packages in each repo. Cache
will be used as long as current repo hash matches the cache.