build: Skip config and transient packages
config and transient packages do not have any code by design, so there's
no need to include them in build. This should fix the problem when
package becomes transient at some point but .a for that package is
present in build folder from an old build and is still included when
linking image.