syscfg: Write defs for each APIs provided by included packages
This adds symbols in syscfg.h for each API provided by packages included
in build. The symbol name has "MYNEWT_API_" prefix followed with sanitized
API name, all illegal characters in resulting name are replaced with "_".

This allows to check if any package included in build provides specific
API. Could be useful if some API is not strictly required, but we can
use it if provided by some package.