Add allowed repositories list
This in addition to ignored repositories list allows to specify
a list of repositories that will be upgraded. The list should be
placed inside project.yml and look for example like this:

project.repositories.allowed:
    - apache-mynewt-core
    - apache-mynewt-nimble
    - apache-mynewt-mcumgr
    - mcuboot
    - arm-CMSIS_5
    - nordic-nrfx
    - mbedtls

If the same repository will be placed on both lists, it will be
ignored by newt.

IMPORTANT NOTE: Ignored repositories will be ignored even if
they were previously downloaded. So if repo was downloaded
by newt upgrade and then added to ignored list,
targets that depend on this repo won't build.