newt: Fix config check crashes
We should use DeepEqual to compare interface values.
In some cases Value interface might contain a map
with some more keys than just "value" key - for
example "description" key. This will cause
DeepEqual to fail, even if the "value" keys
would be the same. However before returning an error
we also check if Expr field of both entries
is not nil. We can assume that in that
case no additional keys in map will be present
and no false positive error will be returned.