newt/build: Workaround for whole-archive flag
Flag whole-archive is not supported on Mac and it was causing
CI failures. Now instead of using this flag with .a file
we just pass a text file with list of all object files from package
to the linker. So now it looks like this:
@pkg.list
instead of:
--whole-archive pkg.a --no-whole-archive

Passing list of object files will result in pulling unreferenced
object code just like in case of whole-archive flag and
should work with all compilers.