builder: Fix false positive flash map errors
To allow conditional overrides of flash map it's necessary to
resolve target, because otherwise the syscfgs that determine
the choice of a different flash map will not be available.
The problem was that every resolving was trying to detect
errors in configuration (also in flash map). As the resolving
had to be done before bsp package reloading, while detecting
errors the default flash map is loaded. If this default flash
map did not define some flash areas used in the target configuration,
an error was reported even if the flash map that should override
the default one was correct.

Now we don't try to detect error while doing this particular
resolving. Instead we do this after reloading bsp package so
the properly chosen flash map is used in error detection.