newt/builder: read BOOTUTIL_MAX_IMG_SECTORS from syscfg for boot trailer
Boot trailer size was hardcoded to 128 sectors, ignoring the value
set in syscfg. This caused mismatches between trailer size and
the actual image size when BOOTUTIL_MAX_IMG_SECTORS was set to
value other than 128, potentially making images not fit
in the slot. Now the trailer size is calculated based on the syscfg
value.